Enable users to reorder events in the day plan of an itinerary
Users can only reorder events within the time of the day. i.e., if they are trying to reorder a morning event, they can only move it within the morning. If they wish to move it to the afternoon/evening, they need to modify the event by clicking on the edit button and change the time of day to anything